+++ src/app/[variants]/(main)/settings/llm/type.ts
@@ -1,5 +1,0 @@
-import { ProviderConfigProps } from './components/ProviderConfig';
-
-export interface ProviderItem extends Omit<ProviderConfigProps, 'id'> {
-  id: string;
-}